+++ neo3/contracts/native/oracle.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 from typing import Optional, cast, List
 from . import NativeContract, register
 from neo3 import contracts, storage, vm
-from neo3.core import types, cryptography, serialization, to_script_hash, msgrouter
+from neo3.core import types, serialization, to_script_hash, msgrouter
 from neo3.network import payloads