+++ neo3/network/payloads/transaction.py
@@ -2,6 +2,8 @@ from __future__ import annotations
 import hashlib
 import abc
 import struct
+import base58  # type: ignore
+import base64
 from enum import Enum
 from typing import List, Optional, Type, TypeVar
 from neo3.core import Size as s, serialization, utils, types, IInteroperable, IJson
@@ -14,6 +16,25 @@ class TransactionAttributeType(Enum):
     HIGH_PRIORITY = 0x1
     ORACLE_RESPONSE = 0x11
 
+    def to_csharp_name(self) -> str:
+        """
+        Internal helper to match C# convention
+        """
+        if self == self.HIGH_PRIORITY:
+            return "HighPriority"
+        else:
+            return "OracleResponse"
+
+    @classmethod
+    def from_csharp_name(cls, name: str):
+        """
+        Internal helper to parse from C# convention
+        """
+        if name == "HighPriority":
+            return cls.HIGH_PRIORITY
+        else:
+            return cls.ORACLE_RESPONSE
+
 
 TransactionAttribute_T = TypeVar('TransactionAttribute_T', bound='TransactionAttribute')
 
@@ -22,6 +43,7 @@ class TransactionAttribute(serialization.ISerializable, IJson):
     """
     Attributes that can be attached to a Transaction.
     """
+
     def __init__(self):
         self.type_: TransactionAttributeType = None
         self.allow_multiple = False
@@ -83,10 +105,12 @@ class TransactionAttribute(serialization.ISerializable, IJson):
         return True
 
     def to_json(self) -> dict:
-        return {"type": self.type_}
+        """ Convert object into json """
+        return {"type": self.type_.to_csharp_name()}
 
     @classmethod
     def from_json(cls, json: dict):
+        """ Create object from JSON """
         c = cls()
         c.type_ = TransactionAttributeType(json["type"])
 
@@ -97,6 +121,12 @@ class HighPriorityAttribute(TransactionAttribute):
         self.type_ = TransactionAttributeType.HIGH_PRIORITY
 
     def verify(self, snapshot: storage.Snapshot, tx: Transaction) -> bool:
+        """
+        Verifies the attribute with the transaction
+
+        Returns:
+            True if verification passes. False otherwise.
+        """
         committee = contracts.NeoToken().get_committee_address(snapshot)
         for signer in tx.signers:
             if signer.account == committee:
@@ -110,7 +140,7 @@ class HighPriorityAttribute(TransactionAttribute):
         pass
 
 
-class Transaction(payloads.IInventory, IInteroperable):
+class Transaction(payloads.IInventory, IInteroperable, IJson):
     """
     Data to be executed by the NEO virtual machine.
     """
@@ -120,6 +150,12 @@ class Transaction(payloads.IInventory, IInteroperable):
     MAX_VALID_UNTIL_BLOCK_INCREMENT = 5760
     #: the maximum number of transaction attributes for a single transaction
     MAX_TRANSACTION_ATTRIBUTES = 16
+
+    HEADER_SIZE = (1  # Version
+                   + 4  # NONCE
+                   + 8  # SYSTEM_FEE
+                   + 8  # NETWORK_FEE
+                   + 4)  # VALID_UNTIL_BLOCK
 
     def __init__(self,
                  version: int,
@@ -132,20 +168,35 @@ class Transaction(payloads.IInventory, IInteroperable):
                  script: bytes = None,
                  witnesses: List[payloads.Witness] = None,
                  protocol_magic: int = None):
+        #: Transaction data structure version - for internal use
         self.version = version
+        #: Random number
         self.nonce = nonce
+        #: The cost of executing the `script`.
         self.system_fee = system_fee
+        #: The cost of validation and inclusion in a block by the consensus node.
         self.network_fee = network_fee
+        #: The maximum chain height this transaction is valid.
         self.valid_until_block = valid_until_block
+        """
+        Optional attributes
+
+        See also:
+            :class:`~neo3.network.payloads.transaction.HighPriorityAttribute`.
+            :class:`~neo3.network.payloads.oracle.OracleResponse`.
+        """
         self.attributes = attributes if attributes else []
         #: A list of authorities used by the :func:`ChecKWitness` smart contract system call.
         self.signers = signers if signers else []
+        #: The array of instructions to be executed on the chain by the virtual machine.
         self.script = script if script else b''
         #: A list of signing authorities used to validate the transaction.
         self.witnesses = witnesses if witnesses else []
 
         # unofficial attributes
+        #: The virtual machine result of executing the `script`.
         self.vm_state = VMState.NONE
+        #: The block height in which the transaction is included.
         self.block_height = 0
         #: The network protocol magic number to use in the Transaction hash function. Defaults to 0x4F454E
         #: Warning: changing this will change the TX hash which can result in dangling transactions in the database as
@@ -199,6 +250,9 @@ class Transaction(payloads.IInventory, IInteroperable):
 
     @property
     def sender(self) -> types.UInt160:
+        """
+        The hash of the account who has send the transaction to the network
+        """
         return self.signers[0].account if len(self.signers) > 0 else types.UInt160.zero()
 
     @property
@@ -219,6 +273,12 @@ class Transaction(payloads.IInventory, IInteroperable):
         writer.write_serializable_list(self.witnesses)
 
     def serialize_unsigned(self, writer: serialization.BinaryWriter) -> None:
+        """
+        Serialize the unsigned part of the object into a binary stream.
+
+        Args:
+            writer: instance.
+        """
         writer.write_uint8(self.version)
         writer.write_uint32(self.nonce)
         writer.write_int64(self.system_fee)
@@ -229,6 +289,7 @@ class Transaction(payloads.IInventory, IInteroperable):
         writer.write_var_bytes(self.script)
 
     def serialize_special(self, writer: serialization.BinaryWriter) -> None:
+        """ Internal use only - serialize the TX includes its unofficial fields """
         self.serialize(writer)
         writer.write_uint8(int(self.vm_state))
         writer.write_uint32(self.block_height)
@@ -246,6 +307,17 @@ class Transaction(payloads.IInventory, IInteroperable):
             raise ValueError("Deserialization error - witness length does not match signers length")
 
     def deserialize_unsigned(self, reader: serialization.BinaryReader) -> None:
+        """
+        Deserialize the unsigned data part of the object from a binary stream.
+
+        Args:
+            reader: instance.
+
+        Raises:
+            ValueError: The version is not zero.
+            ValueError: If the system of network fee is negative.
+            ValueError: If there is no script
+        """
         (self.version,
          self.nonce,
          self.system_fee,
@@ -266,6 +338,7 @@ class Transaction(payloads.IInventory, IInteroperable):
             raise ValueError("Deserialization error - invalid script length 0")
 
     def deserialize_special(self, reader: serialization.BinaryReader) -> None:
+        """ Internal use only - deserialize the data from the stream into a TX that includes the unofficial fields """
         self.deserialize(reader)
         self.vm_state = VMState(reader.read_uint8())
         self.block_height = reader.read_uint32()
@@ -316,11 +389,77 @@ class Transaction(payloads.IInventory, IInteroperable):
         array.append([tx_hash, version, nonce, sender, system_fee, network_fee, valid_until, script])
         return array
 
+    def to_json(self) -> dict:
+        """ Convert object into json """
+        version = b'\x35'
+        # replace with to_address once the feature-wallet branch is merged
+        x = version + self.sender.to_array()
+        sender = base58.b58encode_check(x).decode()
+        return {
+            "hash": "0x" + str(self.hash()),
+            "size": len(self),
+            "version": self.version,
+            "nonce": self.nonce,
+            "sender": sender,
+            "sysfee": str(self.system_fee),
+            "netfee": str(self.network_fee),
+            "validuntilblock": self.valid_until_block,
+            "signers": list(map(lambda s: s.to_json(), self.signers)),
+            "attributes": list(map(lambda a: a.to_json(), self.attributes)),
+            "script": base64.b64encode(self.script).decode(),
+            "witnesses": list(map(lambda w: w.to_json(), self.witnesses))
+        }
+
+    @classmethod
+    def from_json(cls, json: dict, procotol_magic=None):
+        """ Create object from JSON """
+        version = json['version']
+        nonce = json['nonce']
+        system_fee = int(json['sysfee'])
+        network_fee = int(json['netfee'])
+        valid_until_block = json['validuntilblock']
+        attributes: List[TransactionAttribute] = []
+        # ugh :-(
+        for attribute in json['attributes']:
+            try:
+                type_ = payloads.TransactionAttributeType.from_csharp_name(attribute['type'])
+                if type_ == payloads.TransactionAttributeType.HIGH_PRIORITY:
+                    attributes.append(payloads.HighPriorityAttribute())
+                elif type_ == payloads.TransactionAttributeType.ORACLE_RESPONSE:
+                    attributes.append(payloads.OracleResponse.from_json(attribute))
+            except ValueError:
+                raise ValueError("Invalid transaction attribute")
+        signers = list(map(lambda s: payloads.Signer.from_json(s), json['signers']))
+        script = base64.b64decode(json['script'].encode())
+        witnesses = list(map(lambda w: payloads.Witness.from_json(w), json['witnesses']))
+        return cls(version,
+                   nonce,
+                   system_fee,
+                   network_fee,
+                   valid_until_block,
+                   attributes,
+                   signers,
+                   script,
+                   witnesses,
+                   procotol_magic)
+
     def get_script_hashes_for_verifying(self, _: storage.Snapshot) -> List[types.UInt160]:
+        """
+        Helper method to get the data used in verifying the object.
+        """
         return list(map(lambda signer: signer.account, self.signers))
 
     def try_get_attribute(self, needle: Type[TransactionAttribute_T]) -> \
             Optional[TransactionAttribute_T]:
+        """
+        Helper method for finding an attribute of a specific type.
+
+        Args:
+            needle: the type to search for
+
+        Returns:
+            The attribute matching the type if found. None otherwise.
+        """
         for attr in self.attributes:
             if isinstance(attr, needle):
                 return attr